Changed pubDate to be RFC-822 like.
The UNIX-style pubDate was not playing well with Apple Podcast app - the 2-digit year would be shown from 00 (e.g. 0018).  Additionally, the RSS 2.0 spec says pubDate should conform to RFC822 but with a 4-digit year.  As the golang RFC822 format is a 2 digit year, this change creates a RFC822-like pubDate with a 4 digit year.

Also it will remove a dependency.